Added indentation support and one bugfix
Got the indentation working properly and added some support for
printf statements that are on the same line as a for loop.
Doesn't work perfectly yet, you'll have to add the squigly brackets
in around the two statements because now your for loop will have
multiple statements inside of it.

Next up is to make it recognize for loops that just have a printf
statement on the same line and make it add the squigly brackets for
you.